+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/dsl/signer/SignerConfiguration.java
@@ -76,6 +76,7 @@ public class SignerConfiguration {
   private Optional<ClientTlsOptions> downstreamTlsOptions;
   private final Duration startupTimeout;
   private final ChainIdProvider chainIdProvider;
+  private final Optional<KeystoresParameters> v3KeystoresBulkloadParameters;
 
   public SignerConfiguration(
       final String hostname,
@@ -120,7 +121,8 @@ public class SignerConfiguration {
       final int downstreamHttpPort,
       final Optional<ClientTlsOptions> downstreamTlsOptions,
       final ChainIdProvider chainIdProvider,
-      final Optional<String> trustedSetup) {
+      final Optional<String> trustedSetup,
+      final Optional<KeystoresParameters> v3KeystoresBulkloadParameters) {
     this.hostname = hostname;
     this.logLevel = logLevel;
     this.httpRpcPort = httpRpcPort;
@@ -164,6 +166,7 @@ public class SignerConfiguration {
     this.downstreamTlsOptions = downstreamTlsOptions;
     this.chainIdProvider = chainIdProvider;
     this.trustedSetup = trustedSetup;
+    this.v3KeystoresBulkloadParameters = v3KeystoresBulkloadParameters;
   }
 
   public String hostname() {
@@ -344,5 +347,9 @@ public class SignerConfiguration {
 
   public Optional<String> getTrustedSetup() {
     return trustedSetup;
+  }
+
+  public Optional<KeystoresParameters> getV3KeystoresBulkloadParameters() {
+    return v3KeystoresBulkloadParameters;
   }
 }