+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/dsl/signer/SignerConfigurationBuilder.java
@@ -80,6 +80,8 @@ public class SignerConfigurationBuilder {
   private ChainIdProvider chainIdProvider = new ConfigurationChainId(DEFAULT_CHAIN_ID);
   private String trustedSetup;
 
+  private KeystoresParameters v3KeystoresBulkloadParameters;
+
   public SignerConfigurationBuilder withLogLevel(final Level logLevel) {
     this.logLevel = logLevel;
     return this;
@@ -308,6 +310,12 @@ public class SignerConfigurationBuilder {
     return this;
   }
 
+  public SignerConfigurationBuilder withV3KeystoresBulkloadParameters(
+      final KeystoresParameters v3KeystoresBulkloadParameters) {
+    this.v3KeystoresBulkloadParameters = v3KeystoresBulkloadParameters;
+    return this;
+  }
+
   public SignerConfiguration build() {
     if (mode == null) {
       throw new IllegalArgumentException("Mode cannot be null");
@@ -355,6 +363,7 @@ public class SignerConfigurationBuilder {
         downstreamHttpPort,
         Optional.ofNullable(downstreamTlsOptions),
         chainIdProvider,
-        Optional.ofNullable(trustedSetup));
+        Optional.ofNullable(trustedSetup),
+        Optional.ofNullable(v3KeystoresBulkloadParameters));
   }
 }