+++ core/src/main/java/tech/pegasys/web3signer/core/config/Eth1Config.java
@@ -16,6 +16,7 @@ import tech.pegasys.web3signer.core.config.client.ClientTlsOptions;
 import tech.pegasys.web3signer.core.service.jsonrpc.handlers.signing.ChainIdProvider;
 import tech.pegasys.web3signer.signing.config.AwsVaultParameters;
 import tech.pegasys.web3signer.signing.config.AzureKeyVaultParameters;
+import tech.pegasys.web3signer.signing.config.KeystoresParameters;
 
 import java.time.Duration;
 import java.util.Optional;
@@ -47,4 +48,6 @@ public interface Eth1Config {
   AwsVaultParameters getAwsParameters();
 
   long getAwsKmsClientCacheSize();
+
+  KeystoresParameters getV3KeystoresBulkLoadParameters();
 }